Move utils directory into src root.
